API 0.1

+++ app/Http/Controllers/ResidentController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\AddResidentRequest;
 use App\Http\Requests\UpdateResidentRequest;
+use App\Http\Requests\ApiCreateResidentRequest;
+use App\Http\Requests\ApiUpdateResidentRequest;
 use App\Location;
 use App\Resident;
 use App\Scope;
@@ -13,6 +15,7 @@ use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\Support\Facades\Session;
+
 class ResidentController extends Controller
 {
     function __construct()
@@ -339,8 +342,10 @@ class ResidentController extends Controller
                     }
                 }
             }
-            Storage::disk('s3')->delete($resident->image_url);
-            Storage::disk('s3-resized')->delete("resized-".$resident->image_url);
+            if ($resident->image_url != null){
+                Storage::disk('s3')->delete($resident->image_url);
+                Storage::disk('s3-resized')->delete("resized-".$resident->image_url);
+            }
         }
 
         Resident::destroy($ids);
@@ -355,4 +360,241 @@ class ResidentController extends Controller
             ], 200);
         }
     }
+
+
+    //Public APIs
+    /**
+     * Retrieve list of residents from storage.
+     *
+     * @return JSON 
+     */
+    public function getResidents()
+    {
+        $resident = Resident::all();
+        return $resident;
+    }  
+    /**
+     * Retrieve locations with 
+     *
+     * @return JSON 
+     */  
+    public function getRooms()
+    {
+        $rooms = Location::where('location_type_id', 2)->get();
+        return $rooms;
+    }
+    
+    public function createResident(ApiCreateResidentRequest $request)
+    {
+        $resident = Resident::create($request->all());
+        $room = Location::find($request['location_room_id']);
+        $resident->room()->associate($room)->save();
+
+        if(!empty($request['beacon_id'])){
+            $tag = Tag::find((int)$request['beacon_id']);
+            $tag->resident()->save($resident);
+
+            $scopes = Scope::whereHas('policy', function($q){
+                $q->where('deleted_at', null);
+            })->whereNotIn('target_type', ['U', 'C'])->get();
+
+            foreach($scopes as $scope){
+                $scope->tags()->attach($tag);
+                $scope->policy->save();
+            }
+        }
+
+        $image_id = "image-input";
+    
+        if ($request->hasFile($image_id)) {
+            $extension = $request[$image_id]->extension();
+            $filename = "resident-".$resident->resident_id.".".$extension;
+            if ($request->file($image_id)->isValid()) {
+                $validated = $request->validate([
+                    'image-input' => 'mimes:jpeg,png|max:16384',
+                ]);
+                $image_url = Storage::disk('s3')->putFileAs(
+                'residents', $request->file('image-input'), $filename,
+                );
+                $resident->update(['image_url' => $image_url]);
+            }
+        }
+
+        $tagsNull = Tag::doesntHave('resident')
+            ->doesntHave('user')
+            ->pluck('beacon_mac', 'beacon_id');
+        
+        $current = null;
+        if(!empty($resident->tag)){
+            $current = collect([$resident->tag->beacon_id => $resident->tag->beacon_mac]);
+            $tagsNull = $current->concat($tagsNull);
+        }
+
+        $available = true;
+        if($tagsNull->isEmpty()){
+            $available = false;
+        }
+
+        $relationship = Resident::relationship;
+
+        $rooms_ori = Location::where('location_description', 'like', 'Room _')->get();
+        $rooms = [];
+        foreach($rooms_ori as $room){
+            $id = $room->location_master_id;
+            $name = 'L'.$room->floor.' - '.$room->location_description;
+            $rooms[$id] = $name;
+        }
+
+        if ($resident->image_url != null){
+            $resident->image_url = Storage::disk('s3')->url($resident->image_url);
+        }
+
+        return response()->json([
+            "message" => "Resident record created",
+            "resident" => $resident,
+        ], 200);
+        
+    }
+
+    public function updateResident(ApiUpdateResidentRequest $request, $id)
+    {
+        $resident = Resident::where('resident_id', $id)->get()[0];
+        if(!empty($resident->tag)){
+            $pre_tag_id = $resident->tag->beacon_id;
+        }
+        $resident->update($request->all());
+
+        if(!empty($resident->room)){
+            $resident->room()->dissociate()->save();
+        }
+
+        $room = Location::find($request['location_room_id']);
+        $resident->room()->associate($room)->save();
+
+        /** Remove the tag associated with this resident then save new tag (if exist) to resident and scopes associated with this resident*/
+        if(!empty($resident->tag)){ 
+            $resident->tag()->dissociate()->save();
+            $scopes = Scope::whereHas('policy', function($q){
+                $q->where('deleted_at', null);
+            })->whereNotIn('target_type', ['U'])->get();
+
+            if(!empty($request['beacon_id'])){
+                $tag = Tag::find((int)$request['beacon_id']);
+                $tag->resident()->save($resident);
+        
+                foreach($scopes as $scope){
+                    if($scope->target_type == "C"){
+                        if($scope->tags->contains('beacon_id', $pre_tag_id)){
+                            $scope->tags()->detach($pre_tag_id);
+                            $scope->tags()->attach($tag->beacon_id);
+                            $scope->policy->save();
+                        }
+                    } else {
+                        $scope->tags()->detach($pre_tag_id);
+                        $scope->tags()->attach($tag->beacon_id);
+                        $scope->policy->save();
+                    }
+                }
+                
+            } else {
+                foreach($scopes as $scope){
+                    if($scope->target_type == "C"){
+                        if($scope->tags->contains('beacon_id', $pre_tag_id)){
+                            $scope->tags()->detach($pre_tag_id);
+                            $scope->policy->save();
+                        }
+                    } else {
+                        $scope->tags()->detach($pre_tag_id);
+                        $scope->policy->save();
+                    }
+                }
+            }
+        } else {
+            if(!empty($request['beacon_id'])){
+                $tag = Tag::find((int)$request['beacon_id']);
+                $tag->resident()->save($resident);
+
+                $scopes = Scope::whereHas('policy', function($q){
+                    $q->where('deleted_at', null);
+                })->whereIn('target_type', ['A', 'R'])->get();
+    
+                foreach($scopes as $scope){
+                    $scope->tags()->attach($tag->beacon_id);
+                    $scope->policy->save();
+                }
+            }
+        }
+
+        $image_id = "image-input";
+ 
+        if ($request->hasFile($image_id)) {
+            $extension = $request[$image_id]->extension();
+            if ($request->file($image_id)->isValid()) {
+                $validated = $request->validate([
+                    'image-input' => 'mimes:jpeg,png|max:16384',
+                ]);
+                if ($resident->image_url != null){
+                    $image_url = Storage::disk('s3')->delete($resident->image_url);
+                    Storage::disk('s3-resized')->delete("resized-".$resident->image_url);
+                }
+                // return $image_url;
+                $image_url = Storage::disk('s3')->putFile('residents', $request->file('image-input'));
+                $resident->update(['image_url' => $image_url]);
+            }
+        }
+
+        return response()->json([
+            "message" => "Resident record updated",
+            "resident" => $resident,
+        ], 201);
+    }
+
+        /**
+     * Remove the specified resources from storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function destroyResidents($id)
+    {
+        $resident = Resident::find($id);
+        if (!$resident) { 
+            return response()->json([
+                "failure" => "Resident ID not found"
+        ], 200);
+        }
+        if(isset($resident->tag)){
+            $pre_tag_id = $resident->tag->beacon_id;
+            $resident->tag()->dissociate()->save();
+
+            $scopes = Scope::whereHas('policy', function($q){
+                $q->where('deleted_at', null);
+            })->whereNotIn('target_type', ['U'])->get();
+
+            foreach($scopes as $scope){
+                if($scope->target_type == "C"){
+                    if($scope->tags->contains('beacon_id', $pre_tag_id)){
+                        $scope->tags()->detach($pre_tag_id);
+                        $scope->policy->save();
+                    }
+                } else {
+                    $scope->tags()->detach($pre_tag_id);
+                    $scope->policy->save();
+                }
+            }
+        }
+        if ($resident->image_url != null){
+            Storage::disk('s3')->delete($resident->image_url);
+            Storage::disk('s3-resized')->delete("resized-".$resident->image_url);
+        }
+        
+
+        Resident::destroy($id);
+
+        return response()->json([
+                "success" => "Resident deleted successfully."
+        ], 200);
+        
+    }
+
 }
